update passing down props
